clear ram package to cai
