+++ utilities/cai/func_makecompatible.go
@@ -1,0 +1,25 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the 'License');
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an 'AS IS' BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cai
+
+import "strings"
+
+// makeCompatible update a GCP asset ancestryPath to make it compatible with former Policy Library REGO rules
+func makeCompatible(path string) string {
+	path = strings.Replace(path, "organizations", "organization", -1)
+	path = strings.Replace(path, "folders", "folder", -1)
+	path = strings.Replace(path, "projects", "project", -1)
+	return path
+}
